feat: 게시물 수정 삭제시 작성자만 권한을 가지는 로직 추가

+++ src/main/java/com/donation/service/post/PostService.java
@@ -51,8 +51,9 @@ public class PostService {
 
     @Transactional
     public void update(PostUpdateReqDto updateReqDto, LoginMember loginMember, Long id) {
-
-        postRepository.getById(id).changePost(updateReqDto);
+        Post post = postRepository.getById(id);
+        post.validateOwner(loginMember.getId());
+        post.changePost(updateReqDto);
     }
 
     @Transactional
@@ -61,8 +62,9 @@ public class PostService {
     }
 
     @Transactional
-    public void delete(LoginMember loginMember) {
-        Post post = postRepository.getById(loginMember.getId());
+    public void delete(Long id, LoginMember loginMember) {
+        Post post = postRepository.getById(id);
+        post.validateOwner(loginMember.getId());
         postRepository.delete(post);
         commentRepository.deleteByPost(post);
     }
